add useDataState.native.ts, move hook in to its own dir
