+++ packages/react-core/src/hooks/useDataState/types.ts
@@ -1,0 +1,13 @@
+export interface DataState<T> {
+  data: T;
+  hasError: boolean;
+  isLoading: boolean;
+  message: string | undefined;
+}
+
+export type DataAction<T = any, K = any> = (prevData: T, input: K) => T;
+
+export type AsyncDataAction<T = any, K = any> = (
+  prevData: T,
+  input: K
+) => Promise<T>;
