+++ packages/react-core/src/hooks/useDataState/useDataState.native.ts
@@ -1,0 +1,3 @@
+export default function useDataState(): void {
+  throw new Error('useDataState is not implemented for React Native');
+}
